Moved road logic around and disabled it for now. Added positions to simple game obj. Removed Id from barb tribe and stronghold. Should use ObjectId instead. Cleaned up simple game obj creation to always pass in ObjectId. Other refactorings.

+++ Game/Setup/TechnologyFactory.cs
@@ -146,7 +146,8 @@ namespace Game.Setup
             {
                 return new Resource();
             }
-            Resource ret = Ioc.Kernel.Get<StructureCsvFactory>().GetCost(buildType, buildLvl) * 1 / 3;
+
+            Resource ret = Ioc.Kernel.Get<StructureCsvFactory>().GetCost((ushort)buildType, buildLvl) * 1 / 3;
             ret /= 10;
             return ret * 10;
         }
